[Fix] Return bookmarked group only when a group is selected #768
El-Maven encountered crash when no group is selected  and bookmark button (star) is clicked. This was happening because bookmarkPeakGroup function was not handling null pointer case which occurs when no group is selected. This commit will  fix the issue by handling null pointer case and bookmarks group only when user selects a particular group